+++ rust-advent/src/day_three.rs
@@ -5,7 +5,7 @@ pub struct DayThreeSolver {
 }
 
 impl DayThreeSolver {
-    fn tree_encounters(&self, right: usize, down: usize) -> u32 {
+    fn tree_encounters(&self, right: usize, down: usize) -> usize {
         let mut row = 0;
         let mut col = 0;
         let mut tree_count = 0;
@@ -30,14 +30,14 @@ impl Solver for DayThreeSolver {
         let all_lines: Vec<String> = input.lines().map(|l| l.to_string()).collect();
         Ok(Box::new(DayThreeSolver { lines: all_lines }))
     }
-    fn part_one(&self) -> Result<u32, &str> {
+    fn part_one(&self) -> Result<usize, &str> {
         Ok(self.tree_encounters(3, 1))
     }
-    fn part_two(&self) -> Result<u32, &str> {
+    fn part_two(&self) -> Result<usize, &str> {
         let slopes = vec![(1, 1), (3, 1), (5, 1), (7, 1), (1, 2)];
         Ok(slopes
             .iter()
             .map(|slope| self.tree_encounters(slope.0, slope.1))
-            .product::<u32>())
+            .product::<usize>())
     }
 }
